move EvalData to data_utilys.py
